move idempotency logic to package
